Fix addRealMicrosecond() with low float precision

+++ src/Carbon/Traits/Units.php
@@ -41,7 +41,7 @@ trait Units
                 /* @var CarbonInterface $this */
                 $diff = $this->microsecond + $value;
                 $time = $this->getTimestamp();
-                $seconds = floor($diff / static::MICROSECONDS_PER_SECOND);
+                $seconds = (int) floor($diff / static::MICROSECONDS_PER_SECOND);
                 $time += $seconds;
                 $diff -= $seconds * static::MICROSECONDS_PER_SECOND;
                 $microtime = str_pad($diff, 6, '0', STR_PAD_LEFT);
@@ -242,7 +242,7 @@ trait Units
         // Work-around for bug https://bugs.php.net/bug.php?id=75642
         if ($unit === 'micro' || $unit === 'microsecond') {
             $microseconds = $this->micro + $value;
-            $second = floor($microseconds / static::MICROSECONDS_PER_SECOND);
+            $second = (int) floor($microseconds / static::MICROSECONDS_PER_SECOND);
             $microseconds %= static::MICROSECONDS_PER_SECOND;
             if ($microseconds < 0) {
                 $microseconds += static::MICROSECONDS_PER_SECOND;
